meta: Introduce ReadOptions to store interface with Consistency flag
This commit introduces ReadOptions to the store interface
in order to be able to control the level of consistency
for Get operations.

By default, Get operations are linearizable (provided that
the backend offers such guarantee), but a user could place
the Consistency flag at 'false' to disable linearizability
for lower latency requirements.

This only affects Consul and etcd which offer the choice of
linearizability for Get operations.

Signed-off-by: Lukas Prettenthaler <lukas@noenv.com>
Signed-off-by: Alexandre Beslic <abeslic@abronan.com>